use fucking enable_if

+++ AbstractIndexRunner.h
@@ -11,12 +11,12 @@ class AbstractIndexRunner{
 public:
     virtual void initialize() = 0;
     virtual ColumnType getType() = 0;
-    virtual int insert(void * key, size_t value) = 0;
-    virtual size_t findOne(void * key) = 0;
-    virtual size_t findByRange(
+    virtual int insert(void * key, int value) = 0;
+    virtual int findOne(void * key) = 0;
+    virtual int findByRange(
             bool withLeft, void * left, bool leftEqu,
             bool withRight, void * right, bool rightEqu,
-            std::function<void(size_t, size_t)> consumer
+            std::function<void(int, int)> consumer
     ) = 0;
 };
