+++ QueryPlans/SelectQueryPlan.cpp
@@ -17,7 +17,7 @@ SelectQueryPlan::SelectQueryPlan(TableModel * tableModel,
 
 JSON *SelectQueryPlan::runQuery(RecordService *recordService)  {
     auto result = new JSONArray();
-    queryScanner->scan([this, result](size_t id, void *data){
+    queryScanner->scan([this, result](int id, void *data){
         auto current = new JSONArray();
         for(auto & cid : columns){
             auto col = tableModel->getColumn(cid);
