+++ Scanners/QueryScanner.h
@@ -11,12 +11,12 @@
 
 class QueryScanner{
 protected:
-    size_t len;
+    int len;
     RecordService * recordService;
 public:
-    QueryScanner(size_t len, RecordService *recordService);
+    QueryScanner(int len, RecordService *recordService);
     inline virtual ~QueryScanner(){};
-    virtual void scan(std::function<void(size_t, void *)> consumer) = 0;
+    virtual void scan(std::function<void(int, void *)> consumer) = 0;
     virtual JSON * toJSON() = 0;
 };
