+++ Scanners/RangeIndexQueryScanner.h
@@ -17,10 +17,10 @@ class RangeIndexQueryScanner : public QueryScanner{
     bool leq, req;
     bool withLeft, withRight;
 public:
-    RangeIndexQueryScanner(AbstractIndexRunner * indexRunner, RecordService * recordService, size_t len, int tfid,
+    RangeIndexQueryScanner(AbstractIndexRunner * indexRunner, RecordService * recordService, int len, int tfid,
                            void * left, void * right, bool leq, bool req, bool withLeft, bool withRight, std::string on);
 
-    void scan(std::function<void(size_t, void *)> consumer) override ;
+    void scan(std::function<void(int, void *)> consumer) override ;
 
     JSON *toJSON() override ;
 };
