+++ Services/FileService.cpp
@@ -14,35 +14,35 @@ int FileService::openFile(const char * fileName){
     files.emplace_back(fp);
     fseek(fp, 0, SEEK_END);
     int fid = (int)(files.size()) - 1;
-    fileBlockCnt.push_back((size_t)ftell(fp) / BLOCK_SIZE);
+    fileBlockCnt.push_back((int)ftell(fp) / BLOCK_SIZE);
     return fid;
 }
 FILE * FileService::getFile(int fileID){
     assert(fileID < files.size() && fileID >= 0);
     return files[fileID];
 }
-size_t FileService::allocBlock(int fileID){
+int FileService::allocBlock(int fileID){
     assert(fileID < files.size() && fileID >= 0);
     FILE * fp = files[fileID];
     fseek(fp, 0, SEEK_END);
-    size_t offset = (size_t)ftell(fp);
+    int offset = (int)ftell(fp);
     fwrite(zeroBuffer, BLOCK_SIZE, 1,  fp);
     fflush(fp);
     fileBlockCnt[fileID] ++;
     return offset / BLOCK_SIZE;
 }
-void FileService::writeBlock(int fileID, size_t startOffset, void * data){
+void FileService::writeBlock(int fileID, int startOffset, void * data){
     assert(fileID < files.size() && fileID >= 0);
     FILE * fp = files[fileID];
     int a = fseek(fp, startOffset * BLOCK_SIZE, SEEK_SET);
     fwrite(data, BLOCK_SIZE, 1, fp);
     fflush(fp);
 }
-size_t FileService::getBlockCnt(int fileID){
+int FileService::getBlockCnt(int fileID){
     assert(fileID < files.size());
     return fileBlockCnt[fileID];
 }
-void * FileService::readBlock(int fileID, size_t startOffset){
+void * FileService::readBlock(int fileID, int startOffset){
     assert(fileID < files.size());
     FILE * fp = files[fileID];
     fseek(fp, startOffset * BLOCK_SIZE, SEEK_SET);
